[pdu manager] introduce pdu manager (#3013)
Approach
What is the motivation for this PR?
The existing PDU controller implementation cannot handle if DUT connects to 2 different PDUs. The outlet_id was an index in return list, when there is more than one controller, the index will collide. Existing client code uses the outlet_id to access specific outlet. Which will not work in multi-PDU scenario.

How did you do it?
Introduce PDU manager to separate the client from the implementation of pdu controllers.
Change client to toggle outlet with the whole outlet dictionary, so that client don't have to rely on the value of outlet_id.
PDU manager manages all PDU controllers associated with a DUT.
Change devutils to use PDU manager instead of invoking controller directly.
Signed-off-by: Ying Xie ying.xie@microsoft.com