Refactoring the drop_counters test to reuse existing methods (#2998)
Refactoring the existing methods as part of drop-counters testcase to be reused in measuring drops in dualtor testcases.
Move the useful methods already present as part of drop-counter test to helpers dir. This is done so that these methods can now be used by other tests, such as dualtor testcases which require drop measurements.

The present methods are parameterized to receive the packet count for generalized verification.
Tested test_drop_counter on single tor T1 testbed and no regressions were seen. Tested sample test on dualtor.