Implenemt playbook to recover vmhost server automatically (#2800)
This commit introduced several playbook to recover vmhost server automatically.
It's extremely time consuming to redeploy all testbeds on a host server if the server is down or rebooted. This PR adds a new option in testbed-cli.sh to do a cleanup of host server, and adds a respin of vm that failed to start.