[pdu controller] refactoring snmp pdu controller
Making PDU controller more generic. It should manage the whole
PDU controller instead of just for a single DUT.

For PDU models that know has lane concept. Make lane ID part of
the outlet ID. And outlet ID is the partion of OID that can
uniquely identify an outlet on the particular PDU.

Label match is still supported by a passed in hostname.

Add support to query output power wattage

Signed-off-by: Ying Xie <ying.xie@microsoft.com>