[vxlan decap] make sure that count_matched_packets_helper will return (#3049)
Approach
What is the motivation for this PR?
vxlan decap test was observed stuck waiting for matching packet.

How did you do it?
count_matched_packets_helper has a chance to stuck forever if there is always paket received but never have a match.

this change makes sure that the timeout value is honored

Signed-off-by: Ying Xie ying.xie@microsoft.com

How did you verify it?
Run vxlan test on a DUT got pass. and run on the DUT previously stuck, the test fail properly.