Add 'testbed-cli.sh announce-routes'
In PR #2741, test script for announcing IP routes to PTF docker container was deprecated.
The function of announcing IP routes is moved to 'testbed-cli.sh add-topo'. Then we do
not need to run the test script to announce IP routes after a new topology is deployed.

However, there are still ocassions that people may want to re-announce the IP routes to
PTF docker containers. Because announcing routes is part of add-topo, then the only
possible way to re-announce routes is to run 'testbed-cli.sh remove-topo' and
'testbed-cli.sh add-topo'. This is inconvenient and time consuming.

This change added a new 'announce-routes' sub-command to the testbed-cli.sh tool.
We can now just re-announce IP routes using command like below:
    ./testbed-cli.sh -t vtestbed.csv vms-kvm-t0 password.txt

Other changes:
* Corrected the inv_name field in vtestbed.csv file.
* Moved the 'announce_routes.py' module from ansible/roles/vm_set/library to ansible/library

Signed-off-by: Xin Wang <xiwang5@microsoft.com>