multi-asic support for copp testcase (#3060)
This is PR continuation of #2627 where infra changes were done to support copp on multi-asic platforms. In this PR test is enhanced to work on multi-asic platforms. Following are major changes:

Multi-asic as of now will support installing of nanomsg and ptf_nn_agent in syncd directly mode. This way test case will randomly select any front-asic syncd for verification. To support this needed to add iptable rules for HTTP and PTF Traffic forwarding from host to asic namespace. swap_syncd mode is not supported as of now.

All the docker name are translated to particular namespace/asic that is under test. (Added utility api for same).

Remove disable_lldp_for_testing fixture as noted below since test case is not dependent on that.