Adapting fib test for T2 topology
Even though a T2 chassis has multiple DUTs (linecards), we have a single PTF
instance that has all the injected ptf ports that connect to all the DUTs frontpanel ports.
For the fib test, the src port and the expected rcv ports are from this list of all the
injected ptf ports.

In a T2 chassis, we have multiple linecards, and routes learnt from a linecard
are distributed over the fabric to the other linecards. But, the fib on the other
linecards points to the inband recyle port Ethernet-IB0. Therefore, when generating the fib_info_file
for the linecards, if the route is learned over the inband recyle port, we have to figure out which other
linecard this route was learnt on, and use its frontpanel ports as the outgoing ports for this route.

Since a route learnt across the fabric will have an outgoing port on another linecard, and this route
could be learnt from multiple linecards (as is the case with routes announced from T1 VMs in a T2 topology),
the logic to validate the src_mac of the received packet in fib_test/hash_test has been modified to not check
for the src_mac in the expected packet for the verify_packet_any_port call. The src mac is compared to
the target_mac defined in ptf_test_port_map.json for the dut that has the rcvd_port.