[ptfadapter] Check the reachablity of ptf_nn_agent
If two pytest clients try to use the same testbed and connect to the
ptf_nn_agent on the ptf, due to that ptf_nn_agent uses PAIR pattern, the
the second client actually fails to connect to ptf_nn_agent, but
ptfadapter
won't notify the user of this failure.

Add an extra check `_check_ptf_nn_agent_availability` that tries to
connect to the nanomsg socket exposed by ptf_nn_agent, it will verify
if there is one established connection. If not, it will raise an error.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>