Support enabling dynamic buffer after reloading minigraph
`config load_mingraph` is widely used to restore the configuration and
state of the switch during the regression test.
However, after reloading minigraph the dynamic buffer calculation will
be disabled, which means we can not run regression with the feature enabled.
So we need to enable dynamic buffer calculation after reloading minigraph.
The idea is:
- provide a script to enable the dynamic buffer calculation on the fly
- if the dynamic buffer calculation is enabled before reloading minigraph,
  enable it after
- provide an argument to control this behavior

Signed-off-by: Stephen Sun <stephens@nvidia.com>