Use string upto the first '_' to check if this is a supported topology
We have support for DUTs that do not have all its ports connected to a fanout switch (discontiguous ports)
This implies that if we have multiple testbeds (each using different set of ports), then we have to a
topology file per testbed.  For example,
  topo_t1-lag_testbed-1.yml
  topo_t1-lag_testbed-2.yml

The current checks for whether it is a supported topology simply compare the name of the topology (like t1_testbed-1)
to the topologies defined in veos/veos_vtb file to see if it supported. This would require to add
entries for each testbed into the topologies defined in veos/veos_vtb.

Instead we can get use the naming convention that the topology files will use '_' as a delimiter between
the topology we want and the testbed name, then we can use that topology name without the testbed
to check if it is a supported topology. Thus, we only need to put the topology name without the
testbed in topologies of veos/veos_vtb

So, when checking for supported topologies, split the topology name (like t1-lag_testbed-1) with '_'
as a delimiter and checking the first element of the split list as a supported topology
_