[dualtor] Filter out packets destined to PTF backplane interface
We run a PTF script on PTF to sniff traffic. The PTF script calls scapy.sniff which by default capture
on all the PTF interfaces, including the backplane interface for announcing routes from PTF to VMs.
On VMs, the PTF backplane is the next hop for the annoucned routes. So, packets sent by DUT to VMs
are forwarded to the PTF backplane interface as well. Then on PTF, the packets sent by DUT to VMs can
be captured on both the PTF interfaces tapped to VMs and on the backplane interface. This will result in
packet duplication and fail the test. This change is to add capture filter to filter out all the packets
destined to the PTF backplane interface.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>