Increase the PPS_LIMIT_MAX to absord the deviation in slower server case (#3170)
What is the motivation for this PR?
Encounter random test failure on different trap groups with copp policer attached, when it failed the rx_pps is higher than the PPS_LIMIT_MAX.

After investigation found that this is due to the deviation of rx_pps calculation on a slower server.
Although we expect that the PTF should be able to finish all the packets sending in DEFAULT_SEND_INTERVAL_SEC (10s), on a slower server it may need to take more time, so extra DEFAULT_RECEIVE_WAIT_TIME (3s) has already been added.

But we shall also consider the deviation in the rx_pps calculation in this case. As my test, from the TCP dump on DUT, arp request can still be continuously and evenly received on the 11th or 12th second, worst case the 13th second. The calculation of the rx_pps is taking DEFAULT_SEND_INTERVAL_SEC(10s), in this worst scenarion the deviation can be 30%, so the PPS_LIMIT_MAX should be increased from 1.1 to 1.3 times of PPS_LIMIT to cover the worst case.

How did you do it?
Increase the PPS_LIMIT_MAX to PPS_LIMIT * 1.3

How did you verify/test it?
Run copp test cases and it passing

Signed-off-by: Kebo Liu <kebol@nvidia.com>