[dualtor] Filter out packets destined to PTF backplane interface (#3217)
What is the motivation for this PR?
We run a PTF script on PTF to sniff traffic. The PTF script calls scapy's sendrecv.sniff which by default capture
on all the PTF interfaces, including the backplane interface for announcing routes from PTF to VMs.
On VMs, the PTF backplane is the next hop for the announced routes. So, packets sent by DUT to VMs
are forwarded to the PTF backplane interface as well. Then on PTF, the packets sent by DUT to VMs can
be captured on both the PTF interfaces tapped to VMs and on the backplane interface. This will result in
packet duplication and fail the test.

How did you do it?
This change is to add capture filter to filter out all the packets destined to the PTF backplane interface.

Scapy 2.3 supports specifying a list of interfaces to be sniffed on. I have tried scapy 2.3 to sniff only on the
PTF ethx ports. The problem is that v2.3 scapy's sniff function with iface list argument needs a long time
(18 seconds on my test setup, probably different on other test setups) to startup. This makes it difficult
to decide when to start the sender thread after the sniffer thread is started.

After some experiments, it is much more easier to just use the current v2.2 scapy and update the capture
filter to filter out all the traffic destined to the PTF backplane interface.

How did you verify/test it?
Test run a dual tor script to send upstream traffic. Examine the captured pcap file. No packets to the PTF
backplane interface are captured.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>