Parallelizing sanity check and log analyzer on DUTs in a multi-dut tesbed
In a 3 DUT testbed, the sanity check is done once in most of the modules and
took about 3 minutes to execute.  The log analyzer on the 3 DUTs at the end
of every test case took about 10 seconds.

Using the build-in support for parallelization in sonic-mgmt, we see sanity
check on the 3 DUTs to about 90 seconds, and the log analyzer time also cut to
about 7 seconds. These are big reductions in execution time when running lots of
tests in a multi-dut setup