[dualtor][DualTorIO] Fix thread hang waiting for io_ready
For the main thread, it will wait for the packet sender to set event
`io_ready`. If any error occurs within the sender thread and it fails to
set the event, the main thread will hang waiting for the event.

1. Add a error handler to InterruptableThread so it could call this
handelr as callback in case of any errors.

2. Make the sender thread an instance of InterruptableThread and add a
error handler to set `io_ready` to let main thread continue to run.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>