Skip pfc pause lossless test (#3253)
The pfc pause lossless testcase needs fanout to send PFC frames fast enough to completely block the queue. Depending on the link speeds, the inter packet interval to pause the queue varies. We are seeing flakiness in our test runs due to packets getting leaked out of the queue because the frames are not sent fast enough. The reliable way to test this scenario is to use a traffic generator and we have the same tests covered under the ixia folder. Hence skipping this one.

Signed-off-by: Neetha John <nejo@microsoft.com>