Announce route support for T2 chassis topology (#3115)
Approach
What is the motivation for this PR?
Need to advertise routes to the T2 VoQ chassis from the T1 and T3 to simulate a typical deployment of T2 chassis in a data center.

The requirements for the proposed T2 topology in ansible/vars/topo_t2.yml:

Advertise total 12K routes to the chassis
~100 routes from down-strem T1
12K routes from upstream T3
The upstream routes should have a mix of 8, 16 and 24 ECMP paths to the T3 VMs
The down-stream routes should have a mix of 16,32, and 48 ECMP paths across 2 linecards to the T1 VMs
How did you do it?
For T2, we have 3 sets of routes that we are going to advertise

1st set of 1/3 routes are advertised by the first 1/3 of the VMs
2nd set of 1/3 routes are advertised by the remaining 2/3rd of the VMs
3rd set of 1/3 routes are advertised by all the VMs
Also, T1 VM's are distributed over 2 linecards (asics). The same set of routes should be sent by
the same set in both the linecards. So, if linecard1 and linecard2 have T1 VMs connected, then

1st set of routes should be advertised by the first 1/3 of the VMs on linecard1 and also by the first 1/3 of the VMs on linecard2.
2nd set of routes should be advertised by the remaining 2/3 of the VMs on linecard1 and also by the remaining 2/3 of the VMs on linecard2.
3rd set of routes should be advertised by the all VMs on linecard1 and also by all VMs on linecard2
It is assumed that tne number of T1 VMs that on both the linecards is the same.
If we don't have 2 linecards for T1 VMs, then routes above would be advertised only by the first linecard that has T1 VMs
The total number of routes are controlled by the podset_number, tor_number, and tor_subnet_number from the topology file.
With the proposed T2 topology with 400 podsets, and 32 routes per podset we would have 12K routes.
In this topology, we have 24 VMs on each linecard (24 T3 VMs and 48 T1 VMs over 2 linecards).
We would have the following distribution:

T1 Routes:
192.168.xx.xx (32 routes) from the first 8 T1 VM's from linecard2 and linecard3 (VM25-VM32, and VM49-VM56)
192.169.xx.xx (32 routes) from the remaining 16 T1 VM's on linecard2 and linecard3 (VM33-VM48, and VM64-VM72)
192.170.xx.xx (32 routes) from all T1 VMs on linecard2 and linecard3 (VM25-VM48, and VM49-VM72)
T2 Routes:
192.171.xx.xx to 193.45.xx.xx (4K routes) from from first 8 T3 VM's on linecard1 (VM1-VM8)
193.46.xx.xx to 193.176.xx.xx (4K routes) from the remaining 16 T3 VM's on linecard1 (VM9-VM24)
193.177.xx.xx - 194.55.xx.xx (4K routes) from all 24 T3 VM's on linecard1 (VM1-VM24)
default route from all 24 T3 VM's on linecard1 (VM1-VM24)
Other changes:

testbed.py:

Added 't2' as the topology type.
topo_t2.yml:

Updated topo_t2.yml to reflect the changes for routes (podset_number etc.) and also changing the ASN's to align with a typical T2 deployment
Also added some missing template files for ceos dockers when doing add-topo for a T2 topology.

How did you verify/test it?
Ran announce_routes against a T2 chassis