Warmboot KVM tests - replace docker-exec 1s timeout with 5s (#3266)
Increase docker exec timeout from 1s to 5s for KVM warmboot tests:
The timeout for check_docker_exec test is 1s. This timeout is good enough for test in physical devices. However, the KVM devices are inherently slow, and the 1s timeout for check_docker_exec test has intermittently failed in Azure Pipeline PR tests. Therefore, the 1s timeout is increased to 5s for KVM testing. 5s timeout is believed to be generous enough for the KVM device, however more test results are needed to prove this.