Fix incorrect logging of 'is_failed'
In case an ansible module is executed with 'verbose=False', the log message
will not include full result. Instead, the log message will only have some
brief information of the result, including the failure status. The current implementation
uses code like `res.get('is_failed', None)` to get failure indication of ansible module
execution result. Indeed, "is_failed" is not a dict key of the result object. It is an
attribute that is always available. Trying to get "is_failed" key will return None for
most of the results. The fix is to simply get the "is_failed" attribute which is always
available in the result object. Then the log message will have correct failure
indication.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>