Adding code to catch failures in the parallel tasks.
There is an issue parallel_run, where if there is an exception or failure (Exception raised) as part of the target function call,
then it is not captured as a failure, and the process is marked as finished.

As an example, lets consider check_interfaces sanity check. This check calls _find_down_ip_ports which would get the ip interfaces
using show_ip_interface ansible library

  ip_intf_facts = dut.show_ip_interface()['ansible_facts']['ip_interfaces']

On a multi-asic DUT that was running with older sonic code that did not support show ip interface -n <asic#>,
the above show_ip_interface call would fail. However, if this was part of a parallelized taks, then the failure
will be ignored and sanity check would pass.

Added code in parallel.py to check the exitcode of the processes, and if any is non-zero then we would fail.

Also changed the parallel method names in checks.py to know which process started/finished. Before all checks
were called _check_on_dut which did not show process for which sanity check is started/completed.