Announce routes to KVMs and delete arp entry on vm host
Added routes to be announced to the linecards from the eBGP peers.
- ~1K routes to linecard1 from T3 VMs
   - 1/3 from LAG eBGP peer: 192.171.x.x - 192.184.x.x
   - 1/3 from single-port eBGP peer: 192.185.x.x - 192.199.x.x
   - 1/3 from both LAG and single-port eBGP: 192.200.x.x - 192.217.x.x
- ~100 routes to linecard2 from T1 VMs
   - 1/3 from LAG eBGP peer: 192.168.x.x - 192.168.240.x
   - 1/3 from single-port eBGP peer: 192.169.x.x - 192.169.240.x
   - 1/3 from both LAG and single-port eBGP: 192.170.x.x - 192.170.240.x

Also when we start the sonic_vm, we are deleting the arp entry from that VM's mgmt_ip on the server host.
This is required, as otherwise when we run kvmtest.sh twice, the mac address of the VM's mgmt_ip on
server would change in each run, and sometimes in the second run, it would be in 'incomplete' state.

	sudo arp -n | grep 10.250.0.12
	10.250.0.120                     (incomplete)                              br2
	10.250.0.121                     (incomplete)                              br2
	10.250.0.122             ether   52:54:00:c4:a2:f1   C                     br2

This results in the VM mgmt_ip not pingable from the server and thus gives 'AnsibleHostUnreachable' error.