[Keysight] Changes to bind ptf_ip with keysight-api-server container (#3252)
Description of PR
Summary:
Modified add_topo functionalities so that keysight-api-server container can be accessed using corresponding ptf_ip from testbed.csv/yaml file.

Approach
What is the motivation for this PR?
Earlier keysight-api-server container could only be accessed using the testbed-server ip. With this change the api-server can be accessed using the ptf_ip. This makes the operations clear and at per with that of the ptf-containers.

How did you do it?
Added new operation in vm_topology.py to bind only the ptf_ip with the keysight-api-server container.

How did you verify/test it?
Tested with "./testbed-cli add-topo" for both keysight-api-server and ptf-container.