Modifying remove_topo to selectively remove keysight_api_server (#3215)
Modified remove_topo functionalities for Keysight API Server docker container so that by default it will not be removed. It would be removed when a specific flag(viz. remove_keysight_api_server) is used in the cmdline args.

What is the motivation for this PR?
This container is big in size. Since it is not needed to be removed for each remove_topo call an option is provided to remove the container only when it is needed.

How did you do it?
Modified testbed-cli.sh and remove_topo.yml to handle the change.

How did you verify/test it?
Tested on a testbed with keysight API server container already deployed. Ran testbed-cli.sh with remove_topo with and without the flag. It did not impact the remove_topo for ptfs.