[multi-asic] fix everflow ipv6 test (#3326)
While addressing one of comment of PR:#2621 change was made to use self.tx_port_ids instead of EverflowIPv6Tests.tx_port_ids. This change does not work as looks like tx_port_ids intialized in class pytest fixture does not retain the value and become empty list when used within test. This cause failure on multi-asic platforms since if tx_port_ids is empty test case does not work. For Single Asic it was fine since if it is empty we use API get_monitor_port() which is not valid/implemented ulti-asic platforms as of now.