Parallelizing sanity check and log analyzer on DUTs in a multi-dut testbed (#3235)
What is the motivation for this PR?
In a 3 DUT testbed, the sanity check is done once in most of the modules and took about 3 minutes to execute. The log analyzer on the 3 DUTs at the end of every test case took about 10 seconds.

The above adds a lot to the execution time when executing 100's of tests. Need to parallelize the sanity check and log analyzer to reduce the execution time.

How did you do it?
Using the build-in support for parallelization in tests.common.helpers.parallel.parallel_run in sonic-mgmt, we see sanity check on the 3 DUTs to about 90 seconds, and the log analyzer time also cut to about 7 seconds. These are big reductions in execution time when running lots of tests in a multi-dut setup

There is an issue parallel_run, where if there is an exception or failure (Exception raised) as part of the target function call,
then it is not captured as a failure, and the process is marked as finished.

As an example, lets consider check_interfaces sanity check. This check calls _find_down_ip_ports which would get the ip interfaces using show_ip_interface ansible library

 ip_intf_facts = dut.show_ip_interface()['ansible_facts']['ip_interfaces']
On a multi-asic DUT that was running with older sonic code that did not support show ip interface -n <asic#>,the above show_ip_interface call would fail. However, if this was part of a parallelized taks, then the failure will be ignored and sanity check would pass.

Added code in parallel.py to check the exitcode of the processes, and if any is non-zero then we would fail.

This exposed two issues with bgp tests that were using parallel_run that have been fixed as well:

conftest.py - in restore_nbr_gr, the log message at line 130 was using 'k' instead of 'node' and thus failing
test_bgp_bbr.py - in check_other_vms at line 249, we were assiging vm_route as '{'failed': False}'. Thus, we were failing at line 252, where we are checking for key 'vrfs' and that doesn't exist in vm_route. Instead we should have added 'failed' key to vm_route dictionary
How did you verify/test it?
Ran tests in a multi-dut testbed and single testbed to make sure sanity checks are log analysis still happens.