Support for KVM based virtual T2 VoQ chassis (#3218)
Approach
What is the motivation for this PR?
Need to add support for a virtual T2 VoQ chassis using KVMs. This will enable ability to run the sonic-mgmt tests in an virtual environment.

How did you do it?
Detailed explanation of the solution is at docs/testbed/README.testbed.vsChassis.md

A KVM based virtual T2 chassis is a multi-dut setup with 3 single-asic KVM's (2 linecards and 1 supervisor card).
Each linecard has 2 eBGP peers - 1 over a 2-port LAG and 1 over a single port as shown below:

          VM0100                                      VM0101
            ||                                          |
  +---------||------------------------------------------|-------------------+
  |         ||                                          |                   |
  |   +----------------------------------------------------------------+    |
  |   |                         Linecard1                              |    |
  |   +----------------------------------------------------------------+    |
  |                                   |                      |              |
  |   +------------+      +--------------------+     +-----------------+    |
  |   | Supervisor |------|  ovs br-T2Midplane |     | ovs br-T2Inband |    |
  |   +------------+      +--------------------+     +-----------------+    |
  |                                   |                      |              |
  |   +----------------------------------------------------------------+    |
  |   |                          Linecard2                             |    |
  |   +----------------------------------------------------------------+    |
  |           ||                                         |                  |
  +-----------||-----------------------------------------|------------------+
              ||                                         |
	VM0102                                     VM0103
For a T2 chassis, we require the following:

midplane connectivity between all the cards - for voq control path.
inband connectivity between the linecards to forward traffic ingressing one linecard to another linecard.
We add the info above into the topology file (topo_t2-vs.yml) under a new vs_chassis section under DUT

For the inband and midplane connectivity, we are going to use ovs bridges and add the frontpanel ports
defined for the respective functionality to the bridges.
The orchestration of the above topology is done as follows:

Bring up the 3 KVM's as pizza boxes with topology file defined in topo_t2-vs.yml, include deploy-mg.

As part of the bringup:
For midplane connectivity, we create br-T2Midplane and add port eth32 (Ethernet124) of all the KVM's
For inband connectivity, we create br-T2Inband and add port eth31 (Ethernet120) of the linecard KVM's
We then run test_vs_chassis_setup that:

Configures the midplane address on each linecard
The needs to happen even before the config_db is loaded.
We do so by appending linux commands for this are to the end of /etc/rc.local before the last 'exit 0' line.
Add 'chassisdb.conf' to /usr/share/sonic/device/x86_64-kvm_x86_64-r0/ directory.
contents for this on linecards are just the chassis_db_address
contents for this on supervisor card include start_chassis_db.
Since we don't have gen-gm/deploy-mg functionality for t2 voq chassis, we copy hard-coded config_db.json
from tests/vs_voq_cfgs/ to /etc/sonic/config_db.json on each card.
Reboot all the cards.
Test for midplane connectivity by ping the chassis_db_address for all the linecards.
Test for chassis-db connectivity by getting SYSTEM_INTERFACE table using 'redis-cli'
For running the vs chassis, we added 't2' Test group to kvmtest.sh. In this group, we:

skip_sanity and disable_loganalyzer:

Until we have all design PR's merged
the iBGP peers on the inband port don't come up and sanity fails.
some log messages are showing up as errors as well.
Call run_test.sh with test_vs_chassis_setup.py with '-E' option to exit if we have a failure in setting up the vs chassis.

As a sample test, call run_test.sh with test_voq_init.py. In this suite, there are 6 tests, 5 pass and 1 is skipped as we don't have
inband iBGP peers established because of missing code from design.

Some other enhancements:

reboot.py:
Added hostname to all the log/error messages. This was required as we are rebooting all the DUTs in parallel.
announce_routes.py:
For an unsupported topology, don't fail, just log message that announce route is not supported on the topology. This is required until PR #3115 is merged.
changed check for 'is_supervisor_card' to use 'card_type' as the inventory variable, instead of 'type'. For vs, type field is used for the type of vm like 'kvm', 'ceos', 'veos'
How did you verify/test it?
Ran the following

 ./kvmtest.sh -T t2 vms-kvm-t2 vlab-t2-01
Also added routes to be announced to the linecards from the eBGP peers.

~1K routes to linecard1 from T3 VMs (VM0100 and VM0101)
1/3 from LAG eBGP peer: 192.171.x.x - 192.184.x.x
1/3 from single-port eBGP peer: 192.185.x.x - 192.199.x.x
1/3 from both LAG and single-port eBGP: 192.200.x.x - 192.217.x.x
~100 routes to linecard2 from T1 VMs (VM0102 and VM0103)
1/3 from LAG eBGP peer: 192.168.x.x - 192.168.240.x
1/3 from single-port eBGP peer: 192.169.x.x - 192.169.240.x
1/3 from both LAG and single-port eBGP: 192.170.x.x - 192.170.240.x
Also when we start the sonic_vm, we are deleting the arp entry from that VM's mgmt_ip on the server host.
This is required, as otherwise when we run kvmtest.sh twice, the mac address of the VM's mgmt_ip on
server would change in each run, and sometimes in the second run, it would be in 'incomplete' state.

sudo arp -n | grep 10.250.0.12
10.250.0.120                     (incomplete)                              br2
10.250.0.121                     (incomplete)                              br2
10.250.0.122             ether   52:54:00:c4:a2:f1   C                     br2
This results in the VM mgmt_ip not pingable from the server and thus gives 'AnsibleHostUnreachable' error.