Enhance test_sfp to deal with SFP xcvrs which can't reset (#3204)
Approach
What is the motivation for this PR?
SFP xcvrs do not have a reset mechanism: there is no pin and no EEPROM register to perform a reset (according to SFF8472). The existing reset test case checks for a return value of True (i.e. a successful reset), but this would not be possible for SFP xcvrs.

How did you do it?
Added helper function, is_xcvr_resettable, which checks if a xcvr can be reset based on its type. Currently, anything that is not SFP is assumed to be resettable. This function is then called in test_reset test case to determine whether True or False should be the expected return value of calling reset on the xcvr.

How did you verify/test it?
Manually ran test case on a 7260CX3 (which has SFP xcvrs).