[crm] address CRM test issue on multi-dut testbed
CRM test cleanup method always goes to the first DUT, if the testbed
has only one DUT then we never see problem.

But if CRM test is executed on a multi-DUT testbed and DUT other than
first is chosen for the test. Then the cleanup will always be executed
on the first DUT prior to this change.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>