Support for fib test against T2 VoQ chassis with multi-asic linecards (multi-dut)
In a T2 chassis, we have multiple linecards, and routes learnt from a linecard are distributed over the fabric to the other linecards.
But, the fib on the other linecards points to the internal inband recyle port. Therefore, when generating the fib_info_file
we follow the following rules while iterating over all the asics in all the linecards.

- If the nexthop interface for a prefix is an 'internal' port (role in cfg facts is 'int'), then we ignore the route. On at-least one of the linecards
  this prefix would have the nexthop interface as a frontpanel port, and those are the ports that the packet is going to get forwarded to, regardless
  of which linecard the packet ingresses the chassis.  This is similar approach to what is done for the multi-asic support where routes learnt over the
  backplane interfaces are ignored on a asic.
- Treat the whole chassis as a single DUT, ie. generate a single file for all the linecards instead of one for each linecard. This file would have
  all the routes learnt from any of the frontpanel ports on any asic on any linecard, Then, we run the ptf fib test through this file.
   - For this added 'single_fib_for_duts' variable to the ptf fib_test. We need this as the existing ptf_test_port_map and router_macs still are being
     treated as a multi-dut testbed for a VoQ chassis.

Also, since a route learnt across the fabric will have an outgoing port on another linecard, and this route could be learnt from multiple
linecards (as is the case with routes announced from T1 VMs in a T2 topology), the logic to validate the src_mac of the received
packet in fib_test/hash_test has been modified to not check for the src_mac in the expected packet for the verify_packet_any_port
call. The src mac is compared to the target_mac defined in ptf_test_port_map.json for the dut that has the rcvd_port. If it doesn't
match then we fail.