[wr_arp] Fix vlan_port in the response packet (#3438)
test_wr_arp fails for some platforms/images.
This is due to the response packet in Ferret script is generated for the incorrect destination port.
Check if a vxlan_port is configured in the DUT. If yes, use this port while building the packet. Otherwise, use the default value.
Verified in 201911 and master image, and the test_wr_arp passed.