[voq] Neighbor and host IP forwarding test cases and helpers. (#3021)
What is the motivation for this PR?
Implement automated tests to cover the neighbor lifecycle and host IP forwarding sections of the Distributed VoQ Testplan

How did you do it?
Two new test scripts were added, test_voq_nbr.py, and test_voq_ipfwd.py. Additional helpers were added to support these tests, and neighbor validation code was moved from test_voq_init.py to the helper library for code sharing.

test_voq_nbr.py includes tests for clearing the ARP and NDP tables, all at once, and one entry at a time. It also includes tests for MAC address change of the VM with and without unsolicited ARP. Port admin down test is present, and link flap will be added in a future PR. The test script also uses a shared setup fixture to bring down BGP neighbors so the ARP entries aren't immediately relearned after an action, to give the test time to do verifications. The neighbors are brought back up at the end of the script.

test_voq_ipfwd.py implements the "Host IP Forwarding" section of the VoQ test plan. It includes the table validation, link flap and the various positive and negative tests described in the test plan. Not all ports in the T2 topology are tested, there is a function to select 4 ports to use a traffic endpoints and the primary test port is a parameter to all of the test functions.

voq_helpers.py and redis.py were extended to handle neighbor deletes from redis and the Sonic ARP/route tables. Shared fixtures were relocated to conftest.py and the nbrhosts_facts fixture from test_voq_init.py was eliminated, as it was unnecessary.

The existing test_voq_init.py script was updated to use the enum_frontend_dut_hostname fixture for the tests, which eliminate the loops in the tests. The neighbor validation code for test_voq_neighbor_create was moved to voq_helpers so it could be shared with the test_voq_nbr.py script.

How did you verify/test it?
Ran against a chassis with a T2 topology.

Any platform specific information?

Supported testbed topology if it's a new test case?
T2

Co-authored-by: Tom Custodio <thomas.custodio@nokia.com>