[arp]: Cache 'grat_arp' value for dual tor (#3445)
What is the motivation for this PR?
How did you do it?
Prior to changing the 'grat_arp' value on a DUT, store the existing value. During test teardown, if the previous value was 'existing' do not delete the value from the DB.

How did you verify/test it?
Run the test on both dual ToR and single ToR platforms. Confirm that the DB value for grat_arp before and after the test were the same.

Signed-off-by: Lawrence Lee <lawlee@microsoft.com>